requirements.txt
pip
requests
snowflake-connector-python
PyYAML
bs4
regex
datetime
logger
pathlib
